I have coded City entity and Factory.
