+++ src/main/java/za/ac/cput/school_management/factory/lookup/CityFactory.java
@@ -1,4 +1,12 @@
 package za.ac.cput.school_management.factory.lookup;
 
+import za.ac.cput.school_management.domain.lookup.City;
+import za.ac.cput.school_management.domain.lookup.Country;
+
+import java.util.UUID;
+
 public class CityFactory {
+    public static City getCity(String id,String name, Country country){
+        return new City.Builder(id).buildName(name).build();
+    }
 }
